pyproject.toml
[tool.poetry]
name = "editor-ai"
version = "0.1.0"
description = "Gemini 3.0 powered document editor with Greek/English grammar enforcement"
authors = ["EditorAI Team"]
readme = "README.md"
packages = [{include = "core"}]

[tool.poetry.dependencies]
python = "^3.10"
google-genai = "^1.0.0"
streamlit = "^1.40.0"
pydantic = "^2.10.0"
python-docx = "^1.1.2"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

